+++ NEWS.md
@@ -19,6 +19,8 @@ Changes:
     - Added the `generalizability_max_sdm` and `generalizabiltiy_diagnostic` fields to the `cm_diagnostics_summary` table.
     - Added the `mean_before`, `mean_after`, `target_std_diff`, `comparator_std_diff`, and `target_comparator_std_diff` fields to both the `cm_covariate_balance` and `cm_shared_covariate_balance` tables.
     
+7. Improve speed of covariate balance computation.
+    
 
 Bugfixes: